+++ docs/app/documentation/component-docs/localization-editor/localization-editor-header/localization-editor-header.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-localization-editor-header',
+  templateUrl: './localization-editor-header.component.html',
+})
+export class LocalizationEditorHeaderComponent {}